# docs/api/openapi.json
{"openapi":"3.0.1","info":{"title":"Security App API","description":"Backend foundation for the applied cryptography project.","contact":{"name":"MMUD Team"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"securityapp local server"}],"paths":{"/api/tickets/{id}":{"get":{"tags":["ticket-controller"],"operationId":"getTicketById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}}}},"put":{"tags":["ticket-controller"],"operationId":"updateTicket","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketRequest"}}},"required":true},"responses":{"200":{"description":"Ticket updated successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}},"404":{"description":"Ticket or customer not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}}}},"delete":{"tags":["ticket-controller"],"operationId":"deleteTicket","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Ticket not found."},"204":{"description":"Ticket deleted successfully."}}}},"/api/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getCustomerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}},"put":{"tags":["customer-controller"],"operationId":"updateCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequest"}}},"required":true},"responses":{"200":{"description":"Customer updated successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}},"409":{"description":"Customer email already exists.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}},"404":{"description":"Customer not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}},"delete":{"tags":["customer-controller"],"operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Customer deleted successfully."},"404":{"description":"Customer not found."}}}},"/api/tickets":{"get":{"tags":["ticket-controller"],"operationId":"getTickets","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketResponse"}}}}}}},"post":{"tags":["ticket-controller"],"operationId":"createTicket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketRequest"}}},"required":true},"responses":{"201":{"description":"Ticket created successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}},"404":{"description":"Customer not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}}}}},"/api/customers":{"get":{"tags":["customer-controller"],"operationId":"getCustomers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"post":{"tags":["customer-controller"],"operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequest"}}},"required":true},"responses":{"409":{"description":"Customer email already exists.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}},"201":{"description":"Customer created successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"User registered successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"409":{"description":"User email already exists.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"401":{"description":"Invalid email or password.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"200":{"description":"Login successful.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/tickets/{id}/status":{"patch":{"tags":["ticket-controller"],"operationId":"updateTicketStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketStatusUpdateRequest"}}},"required":true},"responses":{"404":{"description":"Ticket not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}},"200":{"description":"Ticket status updated successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}}}}},"/api/health":{"get":{"tags":["system-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"401":{"description":"Authentication required.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"200":{"description":"Current user returned.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/audit-logs":{"get":{"tags":["audit-log-controller"],"operationId":"getAuditLogs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditLogResponse"}}}}}}}},"/api/admin/summary":{"get":{"tags":["admin-controller"],"operationId":"summary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}}},"components":{"schemas":{"TicketRequest":{"required":["createdById","customerId","description","priority","title"],"type":"object","properties":{"customerId":{"type":"integer","format":"int64"},"title":{"maxLength":255,"minLength":0,"type":"string"},"description":{"maxLength":2000,"minLength":0,"type":"string"},"priority":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"status":{"type":"string","enum":["OPEN","PROCESSING","RESOLVED"]},"createdById":{"type":"integer","format":"int64"},"assignedToId":{"type":"integer","format":"int64"}}},"TicketResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int64"},"customerName":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","enum":["OPEN","PROCESSING","RESOLVED"]},"priority":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"createdById":{"type":"integer","format":"int64"},"assignedToId":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CustomerRequest":{"required":["address","email","name","phone","taxCode"],"type":"object","properties":{"name":{"maxLength":150,"minLength":0,"type":"string"},"email":{"maxLength":255,"minLength":0,"type":"string"},"phone":{"maxLength":50,"minLength":0,"type":"string"},"address":{"maxLength":500,"minLength":0,"type":"string"},"taxCode":{"maxLength":50,"minLength":0,"type":"string"}}},"CustomerResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"taxCode":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"required":["email","fullName","password"],"type":"object","properties":{"fullName":{"maxLength":150,"minLength":0,"type":"string"},"email":{"maxLength":255,"minLength":0,"type":"string"},"password":{"maxLength":100,"minLength":8,"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","STAFF","USER"]}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":100,"minLength":8,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"tokenType":{"type":"string"},"role":{"type":"string","enum":["ADMIN","STAFF","USER"]},"user":{"$ref":"#/components/schemas/UserResponse"}}},"TicketStatusUpdateRequest":{"required":["status"],"type":"object","properties":{"status":{"type":"string","enum":["OPEN","PROCESSING","RESOLVED"]}}},"AuditLogResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"action":{"type":"string"},"entityType":{"type":"string"},"entityId":{"type":"integer","format":"int64"},"actorUserId":{"type":"integer","format":"int64"},"actorEmail":{"type":"string"},"success":{"type":"boolean"},"details":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}